// src/types/LoadOrder.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LoadOrderEntry } from "./LoadOrderEntry";

export type LoadOrder = { version: number, baseGameData: string | null, contentOrder?: Array<string> | null, entries: Array<LoadOrderEntry>, };