// src/types/LoadOrderEntry.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PackageKind } from "./PackageKind";
import type { PluginEntry } from "./PluginEntry";

export type LoadOrderEntry = { id: string, name: string, kind: PackageKind, 
/**
 * Path relative to instance data dir, e.g. `Better Bodies`.
 */
relativeDir: string, enabled: boolean, priority: number, plugins: Array<PluginEntry>, treeChecksum: string | null, };