// src/types/Mo2ModlistImportResult.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LoadOrder } from "./LoadOrder";
import type { Mo2ModlistImportReport } from "./Mo2ModlistImportReport";

export type Mo2ModlistImportResult = { loadOrder: LoadOrder, report: Mo2ModlistImportReport, };