// src/types/NerevarManifest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ManifestPackage } from "./ManifestPackage";
import type { RequiredDataFileEntry } from "./RequiredDataFileEntry";
import type { ResolvedOpenMwConfig } from "./ResolvedOpenMwConfig";

export type NerevarManifest = { version: number, instanceId: string, instanceName: string, generatedAt: string, baseGameData: string | null, packages: Array<ManifestPackage>, resolved: ResolvedOpenMwConfig, totalDownloadBytes: number, 
/**
 * TES3MP game server port from the host's `tes3mp-server-default.cfg` [General] section.
 */
tes3mpServerPort: number, tes3mpServerPassword?: string, requiredDataFiles: Array<RequiredDataFileEntry>, };